+++ app/Http/Requests/Movie/ComingSoonMovie/UpdateRequest.php
@@ -17,7 +17,7 @@ class UpdateRequest extends BaseRequest
             'id' => ['required', 'integer', 'exists:coming_soon_movies'],
             'title' => ['required', 'string', "unique:coming_soon_movies,title,{$this->id}"],
             'plot' => ['required', 'string'],
-            'duration_in_minutes' => ['required', 'integer'],
+            'duration_in_minutes' => ['required', 'numeric'],
             'age_restriction' => ['required', 'integer'],
             'country' => ['required', 'string'],
             'language' => ['required', 'string'],
